ステージ5ボス全ての部位の描画を完了 ボスのアームをボス本体から描画するように変更 ボスのアームを左右それぞれに当たり判定が表示されるように処理を追加 EnemyBullet.cppのいらないコメントを削除

+++ ACG/ACG/ObjBullet.cpp
@@ -77,6 +77,16 @@ void CObjBullet::Action()
 	{
 		return;
 	}
+	//ステージ5ボスと当たったら消去
+	if (DeleteCheckObjNameHit(hit, this, OBJ_STAGE5_BOSS))
+	{
+		return;
+	}
+	//ステージ5ボスのアーム(ライトアーム、レフトアーム)と当たったら消去
+	if (DeleteCheckObjNameHit(hit, this, OBJ_STAGE5_BOSS_ARMS))
+	{
+		return;
+	}
 	//岩とあたったら消去
 	if (DeleteCheckObjNameHit(hit, this, OBJ_ROCK))
 	{
